feat(frontend): migrate from create-react-app to Vite
This commit adds support for Vite by updating the package.json file and
creating a new vite.config.js file. The index.html file was also updated
to include new dependencies. The App.jsx file was also updated to remove
the BrowserRouter and replace it with the Router component from
react-router-dom. The index.jsx file was also updated to include the
Provider component from react-redux. Finally, the ProductListScreen.jsx
file was added to the project.