feat(adminUserService.js, UserEditScreen.jsx): add isAdmin field to user object
In adminUserService.js, the updateUser function now updates the isAdmin field of the user object if it is present in the newUser object. In UserEditScreen.jsx, the form now includes a checkbox for the isAdmin field, and the handleChange function updates the user object with the new value of the field.