chore(frontend): update react-router-dom to version 6.6.1 refactor(frontend): replace BrowserRouter with BrowserRouter and Route with Routes in App.jsx refactor(frontend): replace LinkContainer with Link in CheckoutSteps.jsx, Header.jsx, and OrderListScreen.jsx refactor(frontend): replace useHistory with useNavigate in SearchBox.jsx, LoginScreen.jsx, and OrderListScreen.jsx
feat: use react-router-dom's useParams and useNavigate hooks instead of useRouteMatch and useHistory hooks in several screens

This commit replaces the use of useRouteMatch and useHistory hooks with useParams and useNavigate hooks from react-router-dom in several screens. This change was made to improve the performance of the application and to follow best practices.

feat(ProductListScreen.jsx, UserEditScreen.jsx, UserListScreen.jsx): replace react-router-bootstrap with react-router-dom and useHistory with useNavigate hook.